Fix team name validation in UI
Although it _should_ be valid to have an unescaped `-`, at least
Google Chrome and Firefox complain about an "Invalid character in
character class" and completely skip validation, leading to confusion
as to why team creation fails.

Also fixed the validation message to state the minimum number of
characters and applied red text color on invalid input.